chore(MenuProductResponse, ProductRequest, ProductResponse, etc.): change the type of the price attribute of DTOs to BigDecimal, and move MenuGroupRepository into the repository package
