+++ src/main/java/kitchenpos/dto/ProductRequest.java
@@ -1,13 +1,20 @@
 package kitchenpos.dto;
 
+import java.math.BigDecimal;
+
 public class ProductRequest {
     private String name;
-    private Long price;
+    private BigDecimal price;
 
     protected ProductRequest() {
     }
 
-    public ProductRequest(String name, Long price) {
+    public ProductRequest(String name, long price) {
+        this.name = name;
+        this.price = BigDecimal.valueOf(price);
+    }
+
+    public ProductRequest(String name, BigDecimal price) {
         this.name = name;
         this.price = price;
     }
@@ -16,7 +23,7 @@ public class ProductRequest {
         return name;
     }
 
-    public Long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
