+++ src/main/java/kitchenpos/repository/MenuGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.repository;
 
+import kitchenpos.domain.MenuGroupEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroupEntity, Long> {
